refac(cli): Move cli resources to __main__
